ImageDownloader is now smart enough to avoid running duplicate image filters.
It is possible for the same request to get queued up multiple times. We want to call all the completion handlers since we can't tell if they're exact duplicates. However, we want to avoid executing the same filters with the same identifiers over and over. This change ensures that a filter with a given identifier will only be executed once for a request.