Rounded corner radius can now be adjusted by the image scale.
Rounding the corners of an image requires two different approaches. If the image has the same resolution for all screen scales, the radius MUST be divided by the image scale in order to produce the same visual round across all device scales. If the image has multiple resolutions tailored to each screen scale, the same radius can be applied on devices of all screen scales.