Fix for issue #293
Image transition for cached image hit does not need to be called as a delayed async GCD block. It just needs to set the placeholder image before running the image transition. If the delayed block is used, then if the image request is cancelled, the block is not cancelled. And a subsequent new image request that has no image transition and also hits the cache will be possibly overwritten with the delayed block, causing the wrong image to be displayed. This issue is resolved with the code change.